COMP: Cast types

+++ common/MersenneTwisterRNG.cxx
@@ -99,7 +99,7 @@ void
 MersenneTwisterRNG
 ::InitializeUsingTime()
 {
-  unsigned long keys[2] = {time(NULL), clock()};
+  unsigned long keys[2] = {(unsigned long)time(NULL), clock()};
 
   this->InitializeByArray(keys, 2);
 }
